added half of the documentation

+++ lib/panopticon/daemonshepherd/control_socket.py
@@ -1,4 +1,16 @@
 #!/usr/bin/python
+'''
+UNIX and TCP sockets
+--------------------
+
+.. autoclass:: ControlSocket
+   :members:
+
+.. autoclass:: ControlSocketClient
+   :members:
+
+'''
+#-----------------------------------------------------------------------------
 
 import socket
 import os
@@ -7,7 +19,17 @@ import json
 #-----------------------------------------------------------------------------
 
 class ControlSocket:
+  '''
+  Create listening socket, binding to :obj:`address`. If it is a string, UNIX
+  socket is created. For integer or tuple, TCP socket is used. With tuple, the
+  socket is bound to IP address specified as the string part.
+  '''
+
   def __init__(self, address):
+    '''
+    :param address: address to bind to
+    :type address: string, integer or tuple (string, integer)
+    '''
     # address = "path"
     # address = int(port)
     # address = (bindaddr, port)
@@ -29,25 +51,51 @@ class ControlSocket:
     self.close()
 
   def accept(self):
+    '''
+    :rtype: :class:`ControlSocketClient`
+
+    Accept new connection on this socket.
+    '''
     (conn, addr) = self.socket.accept()
     return ControlSocketClient(conn)
 
   def fileno(self):
+    '''
+    Return file descriptor for this socket.
+
+    Method intended for :func:`select.poll`.
+    '''
     return self.socket.fileno()
 
   def close(self):
+    '''
+    Close the socket, possibly removing the file (UNIX socket).
+    '''
     if self.socket is not None:
       self.socket.close()
       self.socket = None
-      os.remove(self.path)
+      if self.path is not None:
+        os.remove(self.path)
 
 #-----------------------------------------------------------------------------
 
 class ControlSocketClient:
+  '''
+  Client socket wrapper for line-based JSON communication.
+  '''
+
   def __init__(self, socket):
+    '''
+    :param socket: connection to client
+    '''
     self.socket = socket
 
   def read(self):
+    '''
+    :rtype: dict, list or scalar
+
+    Read single line of JSON and decode it.
+    '''
     line = self.socket.recv(4096)
     if line == '':
       return None
@@ -60,13 +108,27 @@ class ControlSocketClient:
     except:
       return None
 
-  def send(self, reply):
-    self.socket.send(json.dumps(reply) + "\n")
+  def send(self, message):
+    '''
+    :param message: data structure to serialize as JSON and send to the client
+    :type message: dict, list or scalar
+
+    Send a JSON message to connected client.
+    '''
+    self.socket.send(json.dumps(message) + "\n")
 
   def fileno(self):
+    '''
+    Return file descriptor for this socket.
+
+    Method intended for :func:`select.poll`.
+    '''
     return self.socket.fileno()
 
   def close(self):
+    '''
+    Close the socket.
+    '''
     if self.socket is not None:
       self.socket.close()
       self.socket = None
